cleaned up test code
